Unify service conventions and tidy product dependencies feat/#27
