+++ pyproject.toml
@@ -1,6 +1,8 @@
 [tool.ruff]
 target-version = "py312"
 line-length = 100
+
+[tool.ruff.lint]
 select = [
     "E",   # pycodestyle errors
     "W",   # pycodestyle warnings
@@ -8,6 +10,9 @@ select = [
     "I",   # isort (import 순서 정렬)
     "B",   # flake8-bugbear (잠재적 버그 패턴)
     "UP",  # pyupgrade (Python 최신 문법 권장)
+]
+ignore = [
+    "UP042",  # dev_convention.md 기준: 상태값 Enum은 str + enum.Enum 유지
 ]
 
 [tool.ruff.lint.per-file-ignores]
